Rename lib module to playlist
